+++ LoboHTML/src/main/java/org/loboevolution/html/style/HtmlValues.java
@@ -35,7 +35,7 @@ import org.loboevolution.html.ListValues;
 import org.loboevolution.html.dom.domimpl.HTMLImageElementImpl;
 import org.loboevolution.gui.HtmlPanel;
 import org.loboevolution.html.js.WindowImpl;
-import org.loboevolution.html.node.js.Window;
+import org.loboevolution.js.Window;
 import org.loboevolution.html.renderstate.RenderState;
 import org.loboevolution.info.TimingInfo;
 import org.loboevolution.laf.FontFactory;
@@ -210,7 +210,7 @@ public class HtmlValues {
 				final Font DEFAULT_FONT = FONT_FACTORY.getFont(FontValues.getDefaultFontKey(win.getConfig()));
 				final Font f = (renderState == null) ? DEFAULT_FONT : renderState.getFont();
 				final int fontSize = f.getSize();
-				final double pixelSize = fontSize * dpi / 96;
+				final double pixelSize = (double) (fontSize * dpi) / 96;
 				return (int) Math.round(pixelSize * Double.parseDouble(text));
 			case "rem":
 				final WindowImpl win2 = (WindowImpl) window;
